+++ src/components/OrderItem/OrderItem.module.css
@@ -17,6 +17,11 @@
   box-shadow: 0px 0px 14px 2px rgba(120, 0, 240, 0.3);
 }
 
+.linkOrder {
+  composes: link;
+  height: 246px;
+}
+
 .orderHeaderContainer {
   display: flex;
   justify-content: space-between;
