+++ src/components/ProfileInfo/ProfileInfo.module.css
@@ -1,0 +1,21 @@
+.input {
+  margin-bottom: 24px;
+}
+
+.buttonContainer {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.button {
+  padding: 0;
+  background-color: transparent;
+  border: none;
+  color: #4c4cff;
+  cursor: pointer;
+  transition: all 0.4s linear 0.05s;
+}
+
+.button:disabled {
+  color: #3a3a55;
+}
