+++ src/pages/profile/profile.module.css
@@ -1,7 +1,6 @@
 .main {
   display: flex;
-  width: 100%;
-  margin: 120px 0 0 320px;
+  margin: 40px 0 0 320px;
 }
 
 .menuContainer {
@@ -45,26 +44,4 @@
   padding: 0;
   background: transparent;
   border: none;
-}
-
-.input {
-  margin-bottom: 24px;
-}
-
-.buttonContainer {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.button {
-  padding: 0;
-  background-color: transparent;
-  border: none;
-  color: #4c4cff;
-  cursor: pointer;
-  transition: all 0.4s linear 0.05s;
-}
-
-.button:disabled {
-  color: #3a3a55;
 }
