+++ src/services/reducers/order.ts
@@ -1,13 +1,15 @@
 /* eslint-disable consistent-return */
 /* eslint-disable no-param-reassign */
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
+import { createSlice, createAsyncThunk, PayloadAction } from '@reduxjs/toolkit';
 import { IOrder } from '../../utils/types';
 import api from '../../utils/Api';
+import { isError } from '../../utils/utils';
 
 type TOrderState = {
   orderDetails: IOrder | null;
   orderRequest: boolean;
   orderFailed: boolean;
+  error: string | null;
 };
 
 export const getOrderData = createAsyncThunk<
@@ -15,24 +17,19 @@ export const getOrderData = createAsyncThunk<
   Array<string>,
   { rejectValue: string }
 >('order/get', async (orderIdArray, { rejectWithValue }) => {
-  try {
-    const response = await api.sendOrder(orderIdArray);
+  const response = await api.sendOrder(orderIdArray);
 
-    if (!response.success) {
-      throw new Error('Ошибка получения данных заказа');
-    }
-    return response;
-  } catch (err) {
-    if (err instanceof Error) {
-      return rejectWithValue(err.message);
-    }
+  if (!response.success) {
+    return rejectWithValue(response.message);
   }
+  return response;
 });
 
 const initialState: TOrderState = {
   orderDetails: null,
   orderRequest: false,
   orderFailed: false,
+  error: null,
 };
 
 const orderSlice = createSlice({
@@ -47,15 +44,16 @@ const orderSlice = createSlice({
     builder
       .addCase(getOrderData.pending, (state) => {
         state.orderRequest = true;
+        state.error = null;
         state.orderFailed = false;
       })
       .addCase(getOrderData.fulfilled, (state, action) => {
         state.orderDetails = action.payload;
         state.orderRequest = false;
       })
-      .addCase(getOrderData.rejected, (state, action) => {
+      .addMatcher(isError, (state, action: PayloadAction<string>) => {
+        state.error = action.payload;
         state.orderRequest = false;
-        state.orderFailed = true;
         console.error(action.payload);
       });
   },
