+++ src/utils/utils.ts
@@ -1,0 +1,5 @@
+import { AnyAction } from '@reduxjs/toolkit';
+
+export function isError(action: AnyAction) {
+  return action.type.endsWith('rejected');
+}
